Make Stan data for dynamic coevolutionary model

Make the data list for the Stan model. The function takes a dataset, phylogeny, and a set of variables and performs the necessary computations (e.g., cutting up the tree into segments, computing branch lengths, determining parent and child nodes) for the Stan model.

coev_make_standata(
  data,
  variables,
  id,
  tree,
  effects_mat = NULL,
  complete_cases = FALSE,
  dist_mat = NULL,
  dist_cov = "exp_quad",
  measurement_error = NULL,
  prior = NULL,
  scale = TRUE,
  estimate_Q_offdiag = TRUE,
  estimate_residual = TRUE,
  log_lik = FALSE,
  prior_only = FALSE
)

Arguments

data

An object of class data.frame (or one that can be coerced to that class) containing data of all variables used in the model.

variables

A named list identifying variables that should coevolve in the model and their associated response distributions as character strings (e.g. list(var1 = "bernoulli_logit", var2 = "ordered_logistic")). Must identify at least two variables. Variable names must refer to valid column names in data. Currently, the only supported response distributions are bernoulli_logit, ordered_logistic, poisson_softplus, negative_binomial_softplus, normal, and gamma_log.

id

A character of length one identifying the variable in the data that links rows to tips on the phylogeny. Must refer to a valid column name in the data. The id column must exactly match the tip labels in the phylogeny.

tree

A phylogenetic tree object of class phylo or multiPhylo. The tree(s) must be rooted and must include positive non-zero branch lengths. All trees in multiPhylo objects must have the same number of internal nodes and branches.

effects_mat

(optional) A boolean matrix with row and column names exactly matching the variables declared for the model. If not specified, all cross-lagged effects will be estimated in the model. If specified, the model will only estimate cross-lagged effects where cells in the matrix are TRUE and will ignore cross-lagged effects where cells in the matrix are FALSE. In the matrix, columns represent predictor variables and rows represent outcome variables. All autoregressive effects (e.g., X -> X) must be TRUE in the matrix.

complete_cases

(optional) Logical. If FALSE (default), all missing values are imputed by the model. If TRUE, taxa with missing data are excluded.

dist_mat

(optional) A distance matrix with row and column names exactly matching the tip labels in the phylogeny. If specified, the model will additionally control for spatial location by including a separate Gaussian Process over locations for every coevolving variable in the model.

dist_cov

A string specifying the covariance kernel used for Gaussian Processes over locations. Currently supported are "exp_quad" (exponentiated-quadratic kernel; default), "exponential" (exponential kernel), and "matern32" (Matern 3/2 kernel).

measurement_error

(optional) A named list of coevolving variables and their associated columns in the dataset containing standard errors. Only valid for normally-distributed variables. For example, if we declare variables = list(x = "normal", y = "normal"), then we could set measurement_error = list(x = "x_std_err") to tell the function to include measurement error on x using standard errors from the x_std_err column of the dataset.

prior

(optional) A named list of priors for the model. If not specified, the model uses default priors (see help(coev_fit)). Alternatively, the user can specify a named list of priors. The list must contain non-duplicated entries for any of the following parameters: the autoregressive effects (A_diag), the cross effects (A_offdiag), the Cholesky factor for the drift matrix (L_R), the drift std. dev. parameters (Q_sigma), the continuous time intercepts (b), the ancestral states for the traits (eta_anc), the cutpoints for ordinal variables (c), the overdispersion parameters for negative binomial variables (phi), the shape parameters for gamma variables (shape), the sigma parameters for Gaussian Processes over locations (sigma_dist), the rho parameters for Gaussian Processes over locations (rho_dist), the residual standard deviations when there are repeated observations (sigma_residual), and the Cholesky factor for the residual correlations when there are repeated observations (L_residual). These must be entered with valid prior strings, e.g. list(A_offdiag = "normal(0, 2)").

scale

Logical. If TRUE (default), variables following the normal and gamma_log response distributions are scaled before fitting the model. Continuous variables following the normal distribution are standardised (e.g., mean centered and divided by their standard deviation) and positive real variables following the gamma_log distribution are divided by the mean value without centering. This approach is recommended when using default priors to improve efficiency and ensure accurate inferences. If FALSE, variables are left unscaled for model fitting. In this case, users should take care to set sensible priors on variables.

estimate_Q_offdiag

Logical. If TRUE (default), the model estimates the off-diagonals for the $$Q$$ drift matrix (i.e., correlated drift). If FALSE, the off-diagonals for the $$Q$$ drift matrix are set to zero.

estimate_residual

Logical. If TRUE (default), the model estimates residual standard deviations and residual correlations when there are repeated observations for taxa. If FALSE, residual standard deviations and residual correlations are not estimated. The latter may be preferable in cases where repeated observations are sparse (e.g., only some taxa have only few repeated observations). This argument only applies when repeated observations are present in the data.

log_lik

Logical. Set to FALSE by default. If TRUE, the model returns the pointwise log likelihood, which can be used to calculate WAIC and LOO.

prior_only

Logical. If FALSE (default), the model is fitted to the data and returns a posterior distribution. If TRUE, the model samples from the prior only, ignoring the likelihood.

Value

A list containing the data for fitting the dynamic coevolutionary model in Stan.

Details

For further details, see help(coev_fit)

References

Ringen, E., Martin, J. S., & Jaeggi, A. (2021). Novel phylogenetic methods reveal that resource-use intensification drives the evolution of "complex" societies. EcoEvoRXiv. doi:10.32942/osf.io/wfp95

Sheehan, O., Watts, J., Gray, R. D., Bulbulia, J., Claessens, S., Ringen, E. J., & Atkinson, Q. D. (2023). Coevolution of religious and political authority in Austronesian societies. Nature Human Behaviour, 7(1), 38-45. 10.1038/s41562-022-01471-y

See also

Author

Scott Claessens scott.claessens@gmail.com, Erik Ringen erikjacob.ringen@uzh.ch

Examples

# make stan data
coev_make_standata(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny
  )
#> $N_tips
#> [1] 97
#> 
#> $N_tree
#> [1] 1
#> 
#> $N_obs
#> [1] 97
#> 
#> $J
#> [1] 2
#> 
#> $N_seg
#> [1] 193
#> 
#> $node_seq
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]   98   99  101  192  193  102  111  114  115   116   117   118   100   188
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   119   120   125   126   127   128   156   121   157   178   171   158
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]   122   129   164   165   185   179   172   166   170   180   112   167
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]   174   130   123   103   182   181   124   113   105   159   106   176
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]   131   104   189   151   177   173   184   190   168   183   107   152
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]   160   175   132   133   153   155   134   108   135   139   149   109
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]   186   169   161   146   191   140   162   141   110   136   150   142
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]   187   143   147   144   145   163   148   137     4    95    96   138
#>      [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107] [,108]
#> [1,]    17     70     62     92     53     65     61     44     93     14
#>      [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117] [,118]
#> [1,]     29     66     43     45     42     58     94     40     75     60
#>      [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127] [,128]
#> [1,]     48     88     52      8    154     63     64     79     77     31
#>      [,129] [,130] [,131] [,132] [,133] [,134] [,135] [,136] [,137] [,138]
#> [1,]     51     28     91     82     74     36     47     83     55     69
#>      [,139] [,140] [,141] [,142] [,143] [,144] [,145] [,146] [,147] [,148]
#> [1,]     46     76      5     23     54     32     20     90     25      1
#>      [,149] [,150] [,151] [,152] [,153] [,154] [,155] [,156] [,157] [,158]
#> [1,]      3     50      2     30     15     97     68     38     16     41
#>      [,159] [,160] [,161] [,162] [,163] [,164] [,165] [,166] [,167] [,168]
#> [1,]     22     34     33     59     26     73      7      6     80     37
#>      [,169] [,170] [,171] [,172] [,173] [,174] [,175] [,176] [,177] [,178]
#> [1,]     86     18     81     10     21     35     27     71     84     85
#>      [,179] [,180] [,181] [,182] [,183] [,184] [,185] [,186] [,187] [,188]
#> [1,]     39      9     24     67     49     11     57     56     72     78
#>      [,189] [,190] [,191] [,192] [,193]
#> [1,]     12     13     87     89     19
#> 
#> $parent
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0   98   99   98  192  101  102  111  114   115   116   117    99   118
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]   118   119   120   125   126   127   128   120   156   126   156   157
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]   121   128   158   164   119   178   171   165   157   179   111   166
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]   171   129   122   102   180   180   121   112   103   158   105   174
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]   130   103   114   129   176   172   179   189   167   182   106   151
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]   159   174   131   132   152   153   133   107   134   135   132   108
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]   185   168   160   134   190   139   161   140   108   135   149   141
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]   186   142   146   143   140   162   133   136   104   193   192   137
#>      [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107] [,108]
#> [1,]   116    163    155    191    143    155    151    150    190    113
#>      [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117] [,118]
#> [1,]    170    152    150    149    148    146    189    147    167    130
#>      [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127] [,128]
#> [1,]    145    187    144    109    153    154    154    164    169    175
#>      [,129] [,130] [,131] [,132] [,133] [,134] [,135] [,136] [,137] [,138]
#> [1,]    144    184    191    123    166    173    145    123    138    163
#>      [,139] [,140] [,141] [,142] [,143] [,144] [,145] [,146] [,147] [,148]
#> [1,]    139    168    104    183    138    175    125    117    181    193
#>      [,149] [,150] [,151] [,152] [,153] [,154] [,155] [,156] [,157] [,158]
#> [1,]    101    142    100    170    112    100    161    172    115    148
#>      [,159] [,160] [,161] [,162] [,163] [,164] [,165] [,166] [,167] [,168]
#> [1,]    183    177    177    147    181    165    109    107    127    173
#>      [,169] [,170] [,171] [,172] [,173] [,174] [,175] [,176] [,177] [,178]
#> [1,]    185    188    122    110    178    176    184    162    124    124
#>      [,179] [,180] [,181] [,182] [,183] [,184] [,185] [,186] [,187] [,188]
#> [1,]    131    110    182    160    141    105    136    137    159    169
#>      [,189] [,190] [,191] [,192] [,193]
#> [1,]    106    113    187    186    188
#> 
#> $ts
#>      [,1]       [,2]       [,3]       [,4]        [,5]      [,6]       [,7]
#> [1,]  -99 0.02400105 0.06746978 0.09290211 0.007513379 0.1544719 0.01876492
#>             [,8]        [,9]       [,10]       [,11]       [,12]     [,13]
#> [1,] 0.007647891 0.005921474 0.004474085 0.003112665 0.004550234 0.2668098
#>            [,14]       [,15]      [,16]      [,17]      [,18]       [,19]
#> [1,] 0.007280485 0.008624489 0.01203026 0.01179292 0.01328848 0.001003278
#>            [,20]        [,21]    [,22]       [,23]      [,24]      [,25]
#> [1,] 0.004111867 0.0007227681 0.032275 0.002663366 0.01251495 0.01140792
#>            [,26]      [,27]      [,28]      [,29]       [,30]      [,31]
#> [1,] 0.009187734 0.02877621 0.03426191 0.02296127 0.001496488 0.08805246
#>           [,32]      [,33]      [,34]     [,35]      [,36]     [,37]      [,38]
#> [1,] 0.03865207 0.04334719 0.04463487 0.1026653 0.06633366 0.1961358 0.06357062
#>          [,39]     [,40]     [,41]     [,42]      [,43]      [,44]     [,45]
#> [1,] 0.1499919 0.1304844 0.1446565 0.2737961 0.07381795 0.08128446 0.2085895
#>           [,46]     [,47]     [,48]      [,49]      [,50]      [,51]      [,52]
#> [1,] 0.09694968 0.0381307 0.2084207 0.01451076 0.07404825 0.07257567 0.06386799
#>          [,53]     [,54]      [,55]     [,56]     [,57]      [,58]     [,59]
#> [1,] 0.3184708 0.2200117 0.03150443 0.2128979 0.2268019 0.04408181 0.1672533
#>          [,60]     [,61]      [,62]    [,63]    [,64]     [,65]      [,66]
#> [1,] 0.1365735 0.0954154 0.09996007 0.147063 0.207362 0.1689723 0.01295665
#>           [,67]       [,68]       [,69]    [,70]       [,71]       [,72]
#> [1,] 0.06611392 0.002242045 0.003349449 0.101018 0.005471434 0.003639409
#>           [,73]      [,74]     [,75]     [,76]    [,77]      [,78]    [,79]
#> [1,] 0.03901317 0.02251309 0.4078498 0.1435542 0.104351 0.04982802 0.179864
#>           [,80]      [,81]      [,82]      [,83]      [,84]    [,85]      [,86]
#> [1,] 0.04353043 0.01408332 0.02992327 0.07832502 0.07885602 0.068061 0.01080743
#>           [,87]    [,88]      [,89]       [,90]      [,91]      [,92]     [,93]
#> [1,] 0.06708163 0.012016 0.05588993 0.008448661 0.06148921 0.05073236 0.1211957
#>           [,94]     [,95]    [,96]     [,97]      [,98]     [,99]     [,100]
#> [1,] 0.05780218 0.3307949 0.819174 0.8411389 0.03531966 0.6605537 0.08130539
#>         [,101]   [,102]     [,103]    [,104]    [,105]    [,106]   [,107]
#> [1,] 0.1963106 0.145949 0.09202333 0.1979781 0.3699995 0.1133132 0.334521
#>         [,108]    [,109]    [,110]    [,111]    [,112]     [,113]    [,114]
#> [1,] 0.4120067 0.5232241 0.2750419 0.1170188 0.1852651 0.09123757 0.1596263
#>         [,115]    [,116]    [,117]   [,118]     [,119]    [,120]     [,121]
#> [1,] 0.3824937 0.1039215 0.4870029 0.467679 0.09569461 0.1120389 0.09728522
#>         [,122]    [,123]       [,124]       [,125]    [,126]  [,127]    [,128]
#> [1,] 0.1842164 0.2140188 1.852777e-06 0.0002823631 0.5991136 0.17916 0.2670168
#>         [,129]    [,130]    [,131]    [,132]    [,133]   [,134]    [,135]
#> [1,] 0.1000644 0.3638327 0.1631798 0.4613608 0.5553909 0.369052 0.1010677
#>         [,136]     [,137]    [,138]    [,139]    [,140]    [,141]    [,142]
#> [1,] 0.4624725 0.03811939 0.1011301 0.2070137 0.3271607 0.3973098 0.3172446
#>          [,143]    [,144]    [,145]    [,146]    [,147]    [,148]    [,149]
#> [1,] 0.04126911 0.2720193 0.6599075 0.6972758 0.4483897 0.8830948 0.8922247
#>         [,150]  [,151]    [,152]    [,153]    [,154]   [,155]    [,156]
#> [1,] 0.1264579 0.69307 0.5365641 0.5235932 0.6936258 0.170763 0.5878788
#>         [,157]    [,158]    [,159]    [,160]    [,161]    [,162]    [,163]
#> [1,] 0.7063449 0.1029101 0.3205798 0.3756816 0.3756818 0.1154087 0.4500572
#>         [,164]    [,165]    [,166]    [,167]   [,168]    [,169]    [,170]
#> [1,] 0.6066956 0.1938508 0.3173821 0.6482094 0.375722 0.5982723 0.6878538
#>         [,171]  [,172]    [,173]   [,174]    [,175]    [,176]    [,177]
#> [1,] 0.6134284 0.13878 0.6374391 0.408668 0.3721702 0.1587177 0.4347266
#>         [,178]   [,179]    [,180]    [,181]    [,182]    [,183]    [,184]
#> [1,] 0.4347266 0.408814 0.1404475 0.4603031 0.2793754 0.1426384 0.4313842
#>         [,185]     [,186]    [,187]    [,188]    [,189]    [,190]    [,191]
#> [1,] 0.1425432 0.08492627 0.4292177 0.1960201 0.4215052 0.4360928 0.1368661
#>         [,192]    [,193]
#> [1,] 0.2043183 0.7023055
#> 
#> $tip
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    0    0    0    0    0    0    0    0    0     0     0     0     0     0
#>      [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25] [,26]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,27] [,28] [,29] [,30] [,31] [,32] [,33] [,34] [,35] [,36] [,37] [,38]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,39] [,40] [,41] [,42] [,43] [,44] [,45] [,46] [,47] [,48] [,49] [,50]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,51] [,52] [,53] [,54] [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,63] [,64] [,65] [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,75] [,76] [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86]
#> [1,]     0     0     0     0     0     0     0     0     0     0     0     0
#>      [,87] [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> [1,]     0     0     0     0     0     0     0     0     1     1     1     0
#>      [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107] [,108]
#> [1,]     1      1      1      1      1      1      1      1      1      1
#>      [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117] [,118]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127] [,128]
#> [1,]      1      1      1      1      0      1      1      1      1      1
#>      [,129] [,130] [,131] [,132] [,133] [,134] [,135] [,136] [,137] [,138]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,139] [,140] [,141] [,142] [,143] [,144] [,145] [,146] [,147] [,148]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,149] [,150] [,151] [,152] [,153] [,154] [,155] [,156] [,157] [,158]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,159] [,160] [,161] [,162] [,163] [,164] [,165] [,166] [,167] [,168]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,169] [,170] [,171] [,172] [,173] [,174] [,175] [,176] [,177] [,178]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,179] [,180] [,181] [,182] [,183] [,184] [,185] [,186] [,187] [,188]
#> [1,]      1      1      1      1      1      1      1      1      1      1
#>      [,189] [,190] [,191] [,192] [,193]
#> [1,]      1      1      1      1      1
#> 
#> $effects_mat
#>                     political_authority religious_authority
#> political_authority                   1                   1
#> religious_authority                   1                   1
#> 
#> $num_effects
#> [1] 4
#> 
#> $y
#>       political_authority religious_authority
#>  [1,]                   4                   4
#>  [2,]                   3                   3
#>  [3,]                   3                   3
#>  [4,]                   3                   3
#>  [5,]                   4                   3
#>  [6,]                   1                   1
#>  [7,]                   2                   3
#>  [8,]                   1                   4
#>  [9,]                   4                   3
#> [10,]                   3                   3
#> [11,]                   4                   1
#> [12,]                   2                   3
#> [13,]                   3                   3
#> [14,]                   1                   1
#> [15,]                   4                   4
#> [16,]                   4                   4
#> [17,]                   4                   4
#> [18,]                   4                   4
#> [19,]                   4                   4
#> [20,]                   2                   1
#> [21,]                   3                   3
#> [22,]                   1                   1
#> [23,]                   1                   1
#> [24,]                   3                   3
#> [25,]                   3                   3
#> [26,]                   2                   2
#> [27,]                   3                   3
#> [28,]                   3                   3
#> [29,]                   2                   1
#> [30,]                   2                   2
#> [31,]                   4                   4
#> [32,]                   4                   4
#> [33,]                   3                   1
#> [34,]                   1                   4
#> [35,]                   4                   4
#> [36,]                   4                   4
#> [37,]                   4                   4
#> [38,]                   4                   4
#> [39,]                   4                   4
#> [40,]                   1                   1
#> [41,]                   4                   4
#> [42,]                   4                   4
#> [43,]                   4                   4
#> [44,]                   4                   4
#> [45,]                   3                   3
#> [46,]                   4                   4
#> [47,]                   3                   4
#> [48,]                   4                   4
#> [49,]                   4                   4
#> [50,]                   4                   3
#> [51,]                   4                   4
#> [52,]                   4                   4
#> [53,]                   4                   4
#> [54,]                   3                   3
#> [55,]                   3                   3
#> [56,]                   4                   4
#> [57,]                   4                   4
#> [58,]                   1                   3
#> [59,]                   4                   4
#> [60,]                   4                   4
#> [61,]                   3                   3
#> [62,]                   4                   1
#> [63,]                   4                   4
#> [64,]                   4                   4
#> [65,]                   4                   4
#> [66,]                   4                   4
#> [67,]                   3                   3
#> [68,]                   3                   4
#> [69,]                   1                   2
#> [70,]                   1                   4
#> [71,]                   1                   4
#> [72,]                   3                   3
#> [73,]                   3                   1
#> [74,]                   2                   2
#> [75,]                   1                   3
#> [76,]                   4                   4
#> [77,]                   4                   4
#> [78,]                   3                   3
#> [79,]                   1                   2
#> [80,]                   3                   3
#> [81,]                   4                   4
#> [82,]                   3                   3
#> [83,]                   4                   3
#> [84,]                   4                   4
#> [85,]                   4                   4
#> [86,]                   3                   4
#> [87,]                   1                   4
#> [88,]                   1                   4
#> [89,]                   4                   4
#> [90,]                   3                   3
#> [91,]                   3                   3
#> [92,]                   4                   4
#> [93,]                   4                   4
#> [94,]                   4                   4
#> [95,]                   3                   3
#> [96,]                   4                   4
#> [97,]                   3                   3
#> 
#> $miss
#>       political_authority religious_authority
#>  [1,]                   0                   0
#>  [2,]                   0                   0
#>  [3,]                   0                   0
#>  [4,]                   0                   0
#>  [5,]                   0                   0
#>  [6,]                   0                   0
#>  [7,]                   0                   0
#>  [8,]                   0                   0
#>  [9,]                   0                   0
#> [10,]                   0                   0
#> [11,]                   0                   0
#> [12,]                   0                   0
#> [13,]                   0                   0
#> [14,]                   0                   0
#> [15,]                   0                   0
#> [16,]                   0                   0
#> [17,]                   0                   0
#> [18,]                   0                   0
#> [19,]                   0                   0
#> [20,]                   0                   0
#> [21,]                   0                   0
#> [22,]                   0                   0
#> [23,]                   0                   0
#> [24,]                   0                   0
#> [25,]                   0                   0
#> [26,]                   0                   0
#> [27,]                   0                   0
#> [28,]                   0                   0
#> [29,]                   0                   0
#> [30,]                   0                   0
#> [31,]                   0                   0
#> [32,]                   0                   0
#> [33,]                   0                   0
#> [34,]                   0                   0
#> [35,]                   0                   0
#> [36,]                   0                   0
#> [37,]                   0                   0
#> [38,]                   0                   0
#> [39,]                   0                   0
#> [40,]                   0                   0
#> [41,]                   0                   0
#> [42,]                   0                   0
#> [43,]                   0                   0
#> [44,]                   0                   0
#> [45,]                   0                   0
#> [46,]                   0                   0
#> [47,]                   0                   0
#> [48,]                   0                   0
#> [49,]                   0                   0
#> [50,]                   0                   0
#> [51,]                   0                   0
#> [52,]                   0                   0
#> [53,]                   0                   0
#> [54,]                   0                   0
#> [55,]                   0                   0
#> [56,]                   0                   0
#> [57,]                   0                   0
#> [58,]                   0                   0
#> [59,]                   0                   0
#> [60,]                   0                   0
#> [61,]                   0                   0
#> [62,]                   0                   0
#> [63,]                   0                   0
#> [64,]                   0                   0
#> [65,]                   0                   0
#> [66,]                   0                   0
#> [67,]                   0                   0
#> [68,]                   0                   0
#> [69,]                   0                   0
#> [70,]                   0                   0
#> [71,]                   0                   0
#> [72,]                   0                   0
#> [73,]                   0                   0
#> [74,]                   0                   0
#> [75,]                   0                   0
#> [76,]                   0                   0
#> [77,]                   0                   0
#> [78,]                   0                   0
#> [79,]                   0                   0
#> [80,]                   0                   0
#> [81,]                   0                   0
#> [82,]                   0                   0
#> [83,]                   0                   0
#> [84,]                   0                   0
#> [85,]                   0                   0
#> [86,]                   0                   0
#> [87,]                   0                   0
#> [88,]                   0                   0
#> [89,]                   0                   0
#> [90,]                   0                   0
#> [91,]                   0                   0
#> [92,]                   0                   0
#> [93,]                   0                   0
#> [94,]                   0                   0
#> [95,]                   0                   0
#> [96,]                   0                   0
#> [97,]                   0                   0
#> 
#> $tip_id
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24 25
#> [26] 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50
#> [51] 51 52 53 54 55 56 57 58 59 60 61 62 63 64 65 66 67 68 69 70 71 72 73 74 75
#> [76] 76 77 78 79 80 81 82 83 84 85 86 87 88 89 90 91 92 93 94 95 96 97
#> 
#> $prior_only
#> [1] 0
#>